Fix types
Now that unit tests have same ts config
Split createEditable into own module. Export.
