+++ package.json
@@ -1,4 +1,5 @@
 {
+  "name": "tux",
   "private": true,
   "license": "MIT",
   "scripts": {
@@ -38,7 +39,6 @@
     "@types/jest": "^18.1.1",
     "@types/lodash": "^4.14.61",
     "@types/material-ui": "^0.16.48",
-    "@types/moment": "^2.13.0",
     "@types/mz": "0.0.30",
     "@types/node": "^7.0.4",
     "@types/prop-types": "^15.5.1",
