+++ packages/tux-example-site/package.json
@@ -13,12 +13,12 @@
     "neutrino": "5.0.1",
     "neutrino-preset-tux": "^0.3.0-beta.0",
     "pretty-error": "^2.0.2",
-    "react": "^16.0.0-beta.3",
+    "react": "^16.0.0-beta.5",
     "react-chain": "^0.4.0",
     "react-chain-helmet": "^0.4.0",
     "react-chain-history": "^0.4.0",
     "react-document": "^0.6.0",
-    "react-dom": "^16.0.0-beta.3",
+    "react-dom": "^16.0.0-beta.5",
     "react-helmet": "^5.0.1",
     "source-map-support": "^0.4.14",
     "tux": "^0.3.0-beta.0",
