+++ packages/tux/src/__tests__/globals.ts
@@ -1,3 +1,4 @@
+import 'jest'
 ;(<any>global).requestAnimationFrame = function(callback: () => {}) {
   setTimeout(callback, 0)
 }
