+++ packages/tux/src/components/Editable/Editable.tsx
@@ -1,56 +1,13 @@
-import React, { ReactType, ComponentClass, StatelessComponent } from 'react'
-import PropTypes from 'prop-types'
+import React, { ReactType, ComponentClass, StatelessComponent, ReactNode, ReactElement } from 'react'
 import { EditableProps } from '../../interfaces'
+import { createEditable } from './createEditable'
 
-export interface State {
-  readOnly: boolean,
+export interface Props extends EditableProps {
+  children: ReactNode
 }
 
-export function createEditable<OriginalProps>() {
-  type OuterProps = OriginalProps & EditableProps
-  return function editable(
-    Editor:
-      ComponentClass<OuterProps> |
-      StatelessComponent<OuterProps>
-  ): ComponentClass<OuterProps> {
-    class Editable extends React.Component<OuterProps, State> {
-      static contextTypes = {
-        tux: PropTypes.object,
-        tuxModel: PropTypes.object,
-      }
-
-      static childContextTypes = {
-        tuxModel: PropTypes.object,
-      }
-
-      getChildContext() {
-        return {
-          tuxModel: this.props.model || this.context.tuxModel,
-        }
-      }
-
-      render() {
-        const { model } = this.props
-        const { tux, tuxModel } = this.context
-        return (
-          <Editor
-            {...(this.props as any)}
-            model={model || tuxModel || {}}
-            isEditing={tux.isEditing}
-            tux={tux}
-          />
-        )
-      }
-    }
-
-    return Editable
-  }
-}
-
-const Internal = ({ children }: EditableProps) => (
-  <span>{children}</span>
+const Internal = ({ children }: Props) => (
+  children as ReactElement<any>
 )
 
-export default createEditable()(
-  Internal
-)
+export default createEditable()(Internal)
