+++ packages/tux/src/components/Editable/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Editable'
+export { default as createEditable } from './createEditable'
