+++ packages/tux/src/services/editor.ts
@@ -6,7 +6,7 @@ export function registerEditable(
   type: string,
   value:
     | Array<Field>
-    | ((editorSchema: Map<string, Field>) => Map<string, Field>)
+    | ((editorSchema: Map<string, Field>) => Map<string, Field> | void)
 ) {
   schema.set(type, value)
 }
@@ -40,10 +40,10 @@ export function getEditorSchema(meta: ModelMeta): Array<Field> {
 }
 
 function _mergeSchemas(adapterSchema: Array<Field>, userFunction: Function) {
-  const schemaAsMap = new Map<string, any>(
+  let schemaAsMap = new Map<string, any>(
     adapterSchema.map(field => [field.field, field]) as any
   )
-  userFunction(schemaAsMap)
+  schemaAsMap = userFunction(schemaAsMap) || schemaAsMap
 
   const result = []
   for (const [fieldName, field] of schemaAsMap) {
